pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mcp-solana-ico"
version = "0.0.1"
description = "A simple MCP server for a Solana ICO"
authors = ["waifuai@users.noreply.github.com"]
license = "MIT-0"

[tool.poetry.dependencies]
python = "^3.11"
fastmcp = "^0.4.1" # Update fastmcp to latest
pydantic = "^2.0"
solders = "^0.23.0"
 # Align with solana 0.36.6 requirement
 # Updated to match solana dependency
httpx = "^0.27.0"
 # Allow httpx 0.27+ for mcp compatibility
solana = "^0.36.6" # Update solana to latest
[tool.poetry.group.dev.dependencies]
ruff = "^0.2.0"
pytest = "^7.0"
pytest-asyncio = "^0.21.0"

[tool.ruff]
line-length = 120
select = ["E", "W", "F", "I", "N", "UP", "Y001"]
ignore = ["E501", "F401"]

[tool.ruff.format]
docstring-code-line-length = "source"
